improve runbook output - job description at the start - prefer readable names vs. ids

+++ user/mail/delegate-send-as.ps1
@@ -70,7 +70,15 @@ param
     # CallerName is tracked purely for auditing purposes
     [Parameter(Mandatory = $true)]
     [string] $CallerName
+
 )
+
+if ($Remove) {
+    "## Trying to remove SendAs permission for mailbox '$UserName' from user '$delegateTo'."
+}
+else {
+    "## Trying to give SendAs permission for mailbox '$UserName' to user '$delegateTo'."
+}
 
 try {
     Connect-RjRbExchangeOnline
@@ -79,11 +87,15 @@ try {
     $user = Get-EXOMailbox -Identity $UserName -ErrorAction SilentlyContinue
     if (-not $user) {
         Disconnect-ExchangeOnline -Confirm:$false -ErrorAction SilentlyContinue | Out-Null
-        throw "User $userName has no mailbox."
+        throw "User '$userName' has no mailbox."
     }
 
-    # Just collect trustee for better output
     $trustee = Get-EXOMailbox -Identity $delegateTo -ErrorAction SilentlyContinue
+    # Check if trustee has a mailbox
+    if (-not $trustee) {
+        Disconnect-ExchangeOnline -Confirm:$false -ErrorAction SilentlyContinue | Out-Null
+        throw "Trustee '$delegateTo' has no mailbox."
+    }
 
     if ($Remove) {
         Remove-RecipientPermission -Identity $UserName -Trustee $delegateTo -AccessRights SendAs -confirm:$false | Out-Null
